# Conflicts:
#	src/main/java/com/cg/cart/CartAPI.java
#	src/main/java/com/cg/controller/CustomerController.java
#	src/main/java/com/cg/product/service/ProductServiceImpl.java

+++ src/main/java/com/cg/model/Cart.java
@@ -21,12 +21,11 @@ public class Cart extends BaseEntity {
     @Column(precision = 10, scale = 0, nullable = false)
     private BigDecimal totalAmount;
 
-
     @OneToOne
     @JoinColumn(name = "user_id", nullable = false, foreignKey = @ForeignKey(name = "fk_user_cart"))
     private User user;
 
     @Column(name = "user_id", insertable = false, updatable = false)
-    private int userId;
+    private Long userId;
 
 }
